+++ Runtime/YandexGamesSdk.cs
@@ -1,28 +1,19 @@
+using System;
 using System.Collections;
 using System.Runtime.InteropServices;
+using AOT;
 using UnityEngine;
-#if UNITY_WEBGL && !UNITY_EDITOR
-using UnityEngine.Scripting;
 
-[assembly: AlwaysLinkAssembly]
-#endif
 namespace Agava.YandexGames
 {
     public static class YandexGamesSdk
     {
+        private static Action s_onInitializeSuccessCallback;
+
         /// <summary>
         /// Enable it to log SDK callbacks in the console.
         /// </summary>
         public static bool CallbackLogging = false;
-
-        /// <summary>
-        /// Think of this as a static constructor.
-        /// </summary>
-#if UNITY_WEBGL && !UNITY_EDITOR
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
-#endif
-        [DllImport("__Internal")]
-        private static extern bool YandexGamesSdkInitialize();
 
         /// <summary>
         /// SDK is initialized automatically on load.
@@ -46,12 +37,30 @@ namespace Agava.YandexGames
         private static extern string GetYandexGamesSdkEnvironment();
 
         /// <summary>
-        /// Coroutine waiting for <see cref="IsInitialized"/> to return true.
+        /// Invoke this and wait for coroutine to finish before using any SDK methods.<br/>
+        /// Downloads Yandex SDK script and inserts it into the HTML page.
         /// </summary>
-        public static IEnumerator WaitForInitialization()
+        /// <returns>Coroutine waiting for <see cref="IsInitialized"/> to return true.</returns>
+        public static IEnumerator Initialize(Action onSuccessCallback = null)
         {
+            s_onInitializeSuccessCallback = onSuccessCallback;
+
+            YandexGamesSdkInitialize(OnInitializeSuccessCallback);
+
             while (!IsInitialized)
                 yield return null;
+        }
+
+        [DllImport("__Internal")]
+        private static extern void YandexGamesSdkInitialize(Action successCallback);
+
+        [MonoPInvokeCallback(typeof(Action))]
+        private static void OnInitializeSuccessCallback()
+        {
+            if (CallbackLogging)
+                Debug.Log($"{nameof(YandexGamesSdk)}.{nameof(OnInitializeSuccessCallback)} invoked");
+
+            s_onInitializeSuccessCallback?.Invoke();
         }
     }
 }