+++ Tests/Runtime/DeviceTests.cs
@@ -8,9 +8,10 @@ namespace YandexGames.Tests
     public class DeviceTests
     {
         [UnitySetUp]
-        public IEnumerator WaitForSdkInitialization()
+        public IEnumerator InitializeSdk()
         {
-            yield return YandexGamesSdk.WaitForInitialization();
+            if (!YandexGamesSdk.IsInitialized)
+                yield return YandexGamesSdk.Initialize(SdkTests.TrackSuccessCallback);
         }
 
         [Test]