+++ Tests/Runtime/SdkTests.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using Agava.YandexGames;
 using NUnit.Framework;
@@ -7,16 +8,30 @@ namespace YandexGames.Tests
 {
     public class SdkTests
     {
+        private static bool s_isInitializationSuccessCallbackReceived = false;
+
         [UnitySetUp]
-        public IEnumerator WaitForSdkInitialization()
+        public IEnumerator InitializeSdk()
         {
-            yield return YandexGamesSdk.WaitForInitialization();
+            if (!YandexGamesSdk.IsInitialized)
+                yield return YandexGamesSdk.Initialize(TrackSuccessCallback);
         }
 
         [Test]
         public void ShouldReturnEnvironment()
         {
             Assert.IsNotEmpty(YandexGamesSdk.Environment.browser.lang);
+        }
+
+        [Test]
+        public void ShouldReceiveInitializationSuccessCallback()
+        {
+            Assert.IsTrue(s_isInitializationSuccessCallbackReceived);
+        }
+
+        public static void TrackSuccessCallback()
+        {
+            s_isInitializationSuccessCallbackReceived = true;
         }
     }
 }